Faster nesting algorithm
